AQ-335: Change the information being passed to the custom graph, add enums for timeframe
