+++ src/constants/graphEnums.js
@@ -1,0 +1,6 @@
+export const timeframeEnums = {
+  DAY: "day",
+  FIVE: "five",
+  MONTH: "month",
+  YEAR: "year",
+};
